[MOD] stock 연관 테이블 type 수정, BaseEntity 추가 #15

+++ src/main/java/com/synergyx/trading/repository/StockDetailRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StockDetailRepository extends JpaRepository<StockDetail, Integer> {
+public interface StockDetailRepository extends JpaRepository<StockDetail, Long> {
 }
